// src/API.ts
export type CreateProductInput = {
  id?: string | null,
  name: string,
  price: number,
  weight?: number | null,
  options: Array< ProductOptionsInput >,
  thumbnail: string,
  images: Array< string >,
  description: string,
  avilable: boolean,
  unlimited: boolean,
  extraDetails?: string | null,
};

export type ProductOptionsInput = {
  name: string,
  thumbnail: string,
  colorCode: string,
  stock: number,
};